put the Enum proc macro in the existing macro crate!
