+++ macro/Cargo.toml
@@ -30,3 +30,4 @@ sqlx = "0.8.2"
 lazy_static = "1"
 once_cell = "1"
 itertools = "0.13.0"
+heck = "0.5"
