--- macro_derive_enum/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "ormlite-derive-enum"
-version = "0.22.7"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1.0"
-syn = { version = "2.0", features = ["full"] }
-proc-macro2 = "1.0"
