+++ ormlite/Cargo.toml
@@ -58,7 +58,6 @@ sqlx = { version = "0.8.2" }
 tokio = { version = "1.40.0", features = ["full"] }
 ormlite-macro = { path = "../macro" , version = "0.22.2" }
 ormlite-core = { path = "../core" , version = "0.22.2" }
-ormlite-derive-enum = { path = "../macro_derive_enum", version = "0.22.2" }
 sqlx-core = "0.8.2"
 sqlmo = "0.22"
 tokio-stream = "0.1.16"
